requirements.txt
# B.O.N.S.A.I. Agent Economy API
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Vector search
pinecone==5.4.2

# LLM
anthropic==0.43.0

# x402 payment verification
eth-account==0.13.4
web3==7.6.0

# Rate limiting
slowapi==0.1.9

# Observability
structlog==24.4.0

# Embeddings (OpenAI text-embedding-3-small)
openai==1.59.6

# Utils
httpx==0.28.1
python-dotenv==1.0.1
